# api/openapi/marketplace.v1.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Публичный REST API маркетплейса. Контракт является источником правды (contract-first).
servers:
  - url: https://api.marketplace.kz
  - url: http://localhost:8080
...
paths:
...
  /api/v1/orders:
    post:
      tags: [orders]
      operationId: placeOrder
      description: "Оформляет заказ из корзины: резерв остатка, фиксация цен и промокода, создание платежа. Ответ содержит ссылку на страницу оплаты."
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlaceOrderRequest" }
      responses:
        "201":
          description: Заказ ожидает оплаты
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlacedOrder" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    get:
      tags: [orders]
      operationId: listOrders
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/OrderStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Заказы покупателя, новые первыми
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/OrderSummary" }
                  page: { $ref: "#/components/schemas/PageInfo" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/orders/{id}:
    get:
      tags: [orders]
      operationId: getOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200": { $ref: "#/components/responses/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orders/{id}/cancel:
    post:
      tags: [orders]
      operationId: cancelOrder
      description: "Отмена покупателем или поддержкой до передачи в доставку; после оплаты выполняется возврат средств."
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reason" }
      responses:
        "200": { $ref: "#/components/responses/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/payments/{id}:
    get:
      tags: [payments]
      operationId: getPayment
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Платёж с возвратами
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/payments/{id}/retry:
    post:
      tags: [payments]
      operationId: retryPayment
      description: Новая попытка оплаты заказа после отказа, пока не истёк срок оплаты.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Ссылка на оплату
          content:
            application/json:
              schema:
                type: object
                required: [order_id, payment_id, payment_url]
                properties:
                  order_id: { type: string, format: uuid }
                  payment_id: { type: string, format: uuid }
                  payment_url: { type: string, format: uri }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/me/payment-methods:
    get:
      tags: [payments]
      operationId: listPaymentMethods
      description: Сохранённые способы оплаты; хранятся только токены провайдера.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Способы оплаты
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, provider, label, created_at]
                      properties:
                        id: { type: string, format: uuid }
                        provider: { type: string }
                        label: { type: string }
                        created_at: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/me/payment-methods/{id}:
    delete:
      tags: [payments]
      operationId: removePaymentMethod
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "204": { description: Способ оплаты удалён }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/seller/orders:
    get:
      tags: [seller-orders]
      operationId: listSellerOrders
      description: Оплаченные и отменённые заказы с позициями продавца, без адреса покупателя.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: seller_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/OrderStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Заказы продавца
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SellerOrder" }
                  page: { $ref: "#/components/schemas/PageInfo" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/seller/orders/{id}/ship:
    post:
      tags: [seller-orders]
      operationId: shipOrder
      description: "Продавец отмечает заказ переданным в доставку (paid/in_fulfilment → shipped). Этап 4: ручная отметка вместо интеграции с логистикой (ADR-0018)."
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "204": { description: Заказ помечен как переданный в доставку }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/admin/checkout-sagas:
    get:
      tags: [support]
      operationId: listCheckoutSagas
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [running, compensating, completed, compensated, manual] }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Саги оформления, недавно изменённые первыми
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/CheckoutSaga" }
                  page: { $ref: "#/components/schemas/PageInfo" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/checkout-sagas/{id}/resume:
    post:
      tags: [support]
      operationId: resumeCheckoutSaga
      description: Возобновляет компенсации саги в статусе manual.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "204": { description: Компенсации выполнены }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/admin/payments/reconciliations/{day}:
    get:
      tags: [support]
      operationId: getReconciliation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: day
          in: path
          required: true
          schema: { type: string, format: date }
        - name: provider
          in: query
          schema: { type: string, default: sandbox }
      responses:
        "200":
          description: Отчёт сверки с провайдером
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reconciliation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /webhooks/payments/{provider}:
    post:
      tags: [webhooks]
      operationId: receivePaymentWebhook
      description: "Вебхук провайдера: подпись в заголовке провайдера (для эмулятора X-Sandbox-Signature: t=<unix>,v1=<hex>), окно ±5 минут, allowlist адресов, тело до 64 КБ."
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Событие принято
          content:
            application/json:
              schema:
                type: object
                required: [received, duplicate]
                properties:
                  received: { type: boolean }
                  duplicate: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderStatus:
      name: status
      in: query
      schema:
        type: string
        enum: [created, awaiting_payment, paid, in_fulfilment, shipped, delivered, completed, cancelled, failed, returning, returned]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Order:
      description: Заказ
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Order" }
...
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Операция запрещена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ресурс не найден
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Конфликт состояния
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnprocessableEntity:
      description: Нарушено бизнес-правило
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: string }
        has_more: { type: boolean }
    Reason:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
...
    Address:
      type: object
      required: [recipient, phone, city, line]
      properties:
        recipient: { type: string, maxLength: 200 }
        phone: { type: string, description: 10–15 цифр, допускаются пробелы, скобки и дефисы }
        country: { type: string, default: KZ, minLength: 2, maxLength: 2 }
        city: { type: string, maxLength: 100 }
        line: { type: string, maxLength: 300 }
        postal_code: { type: string, maxLength: 20 }
    PlaceOrderRequest:
      type: object
      required: [address]
      properties:
        address: { $ref: "#/components/schemas/Address" }
        delivery_method: { type: string, enum: [standard], default: standard }
        payment:
          type: object
          properties:
            method_id: { type: string, format: uuid, description: Сохранённый способ оплаты }
            save_method: { type: boolean, description: Сохранить токен способа оплаты после успешной оплаты }
        expected_total: { type: integer, format: int64, description: Итог из предпросмотра; при расхождении — ORDER_TOTAL_CHANGED }
    PlacedOrder:
      type: object
      required: [order_id, status, payment_id, payment_url, total, currency]
      properties:
        order_id: { type: string, format: uuid }
        status: { type: string, enum: [awaiting_payment] }
        payment_id: { type: string, format: uuid }
        payment_url: { type: string, format: uri }
        total: { type: integer, format: int64 }
        currency: { type: string }
    OrderItem:
      type: object
      required: [sku, product_id, seller_id, title, quantity, unit_price, subtotal, discount, total]
      properties:
        sku: { type: string }
        product_id: { type: string }
        seller_id: { type: string, format: uuid }
        title: { type: string }
        quantity: { type: integer }
        unit_price: { type: integer, format: int64 }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
    OrderPart:
      type: object
      required: [seller_id, subtotal, discount, shipping, total]
      properties:
        seller_id: { type: string, format: uuid }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        shipping: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
    Order:
      type: object
      required: [id, status, address, delivery_method, currency, subtotal, discount, shipping, total, items, parts, history, cancellable, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        address: { $ref: "#/components/schemas/Address" }
        delivery_method: { type: string }
        promo_code: { type: string }
        currency: { type: string }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        shipping: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        parts:
          type: array
          items: { $ref: "#/components/schemas/OrderPart" }
        history:
          type: array
          items:
            type: object
            required: [to, actor, at]
            properties:
              from: { type: string }
              to: { type: string }
              actor: { type: string, enum: [buyer, seller, support, system] }
              reason: { type: string }
              at: { type: string, format: date-time }
        payment:
          type: object
          required: [payment_id, status]
          properties:
            payment_id: { type: string, format: uuid }
            status: { type: string, enum: [created, pending, authorized, captured, failed, cancelled, refunded] }
            redirect_url: { type: string, format: uri }
            pay_before: { type: string, format: date-time }
        cancellable: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderSummary:
      type: object
      required: [id, status, currency, total, items_count, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        currency: { type: string }
        total: { type: integer, format: int64 }
        items_count: { type: integer }
        created_at: { type: string, format: date-time }
    SellerOrder:
      type: object
      required: [id, status, currency, items, part, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        currency: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        part: { $ref: "#/components/schemas/OrderPart" }
        created_at: { type: string, format: date-time }
    CheckoutSaga:
      type: object
      required: [order_id, buyer_id, status, step, reservation_id, payment_id, attempts, compensated, deadline, updated_at]
      properties:
        order_id: { type: string, format: uuid }
        buyer_id: { type: string, format: uuid }
        status: { type: string, enum: [running, compensating, completed, compensated, manual] }
        step: { type: string }
        reservation_id: { type: string }
        payment_id: { type: string }
        reason: { type: string }
        last_error: { type: string }
        attempts: { type: integer }
        compensated:
          type: array
          items: { type: string, enum: [cancel_payment, refund_payment, return_stock, release_promo, close_order] }
        deadline: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Payment:
      type: object
      required: [id, order_id, provider, status, currency, amount, authorized, captured, refunded, refunds, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        provider: { type: string }
        status: { type: string, enum: [created, pending, authorized, captured, failed, cancelled, refunded] }
        currency: { type: string }
        amount: { type: integer, format: int64 }
        authorized: { type: integer, format: int64 }
        captured: { type: integer, format: int64 }
        refunded: { type: integer, format: int64 }
        redirect_url: { type: string, format: uri }
        failure_reason: { type: string }
        refunds:
          type: array
          items:
            type: object
            required: [id, amount, status, created_at]
            properties:
              id: { type: string, format: uuid }
              amount: { type: integer, format: int64 }
              reason: { type: string }
              status: { type: string, enum: [pending, succeeded, failed] }
              created_at: { type: string, format: date-time }
              completed_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Reconciliation:
      type: object
      required: [day, provider, checked, mismatches, created_at]
      properties:
        day: { type: string, format: date }
        provider: { type: string }
        checked: { type: integer }
        mismatches:
          type: array
          items:
            type: object
            required: [provider_payment_id, field, internal, provider]
            properties:
              payment_id: { type: string, format: uuid }
              provider_payment_id: { type: string }
              field: { type: string, enum: [presence, captured, refunded] }
              internal: { type: string }
              provider: { type: string }
        created_at: { type: string, format: date-time }
...